Remove redundant check
If the pathname is "." it also starts with ".", so we do not need to
check this here.

Also, if it is ".." it also starts with ".", so we do not need to check
this either.

One step further, we don't even need this function anymore, because we
only alias another function here.