Simplify and fix logic for hidden files
When checking if a path should be hidden or not, we used the file_name.
On the first iteration of iterate_folders, the full path was sent. e.g.
    "../path/to/.file".
Instead of doing a complicated check here, we always allow the root path
(levels.len() == 0). That way we only have to check whether the path
starts with a ".".

This fixes the bug where "tree-rs ../.." showed no content.