Remove BufWriter due to Windows compatibility
The BufWriter threw a panic when unwrapped on Windows. I've tried to
research why, but couldn't find anything. This will regrettably decrease
performance, but as Windows is my main motivation, I think it's ok.

An alternative is to use the terminfo crate which appears to have a
buffered colored printer that also works on Windows.